api-spec-gen/requirements.txt
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
openapi-spec-validator==0.7.1
pyyaml==6.0.2
requests==2.32.3
openai==1.43.0
anthropic==0.34.0
